Code Cleanup: * Move IntPair to structs package * Update PrimeNumberTools to use Factoring instead of deprecated (removed) NumberFactors method
